+++ internal/websocket.go
@@ -5,6 +5,7 @@ import (
 	"log/slog"
 	"net/http"
 	"strings"
+	"time"
 
 	"slices"
 
@@ -46,33 +47,43 @@ func (w *WebSocketServer) hardRefresh(conn *websocket.Conn) {
 	// w.clientsLock.Lock()
 	// w.streamLock.Lock()
 	// w.transcriptLock.Lock()
-
 	outData := HardRefreshData{
 		Event: "hardrefresh",
 		Data:  w.clientData,
 	}
+	startTime := time.Now()
+	MessagesTotal.Inc()
 	if err := conn.WriteJSON(outData); err != nil {
+		WebsocketError.Inc()
 		defer w.closeSocket(conn)
 	}
 
 	// w.transcriptLock.Unlock()
 	// w.streamLock.Unlock()
 	// w.clientsLock.Unlock()
+	MessageProcessingDuration.Observe(time.Since(startTime).Seconds())
 }
 
 func (w *WebSocketServer) broadcast(msg []byte) {
+	startTime := time.Now()
+	MessageSize.Observe(float64(len(msg)))
+	MessagesTotal.Inc()
 	w.clientsLock.Lock()
 	for _, c := range w.clients {
 		go func(msg []byte) {
 			if err := c.WriteMessage(websocket.TextMessage, msg); err != nil {
+				WebsocketError.Inc()
 				defer w.closeSocket(c)
 			}
 		}(msg)
 	}
 	w.clientsLock.Unlock()
+	MessageProcessingDuration.Observe(time.Since(startTime).Seconds())
 }
 
 func (w *WebSocketServer) closeSocket(conn *websocket.Conn) error {
+	ActiveConnections.Dec()
+	ClientsPerKey.WithLabelValues(w.key).Dec()
 
 	w.clientsLock.Lock()
 	for i, c := range w.clients {
@@ -90,6 +101,7 @@ func (ws *WebSocketServer) wsHandler(w http.ResponseWriter, r *http.Request) {
 	if ws.clientConnections >= ws.maxConn {
 		http.Error(w, "Max number of connection already reached", http.StatusBadRequest)
 		slog.Error("max number of connections already reached", "key", ws.key, "func", "wsHandler", "maxConn", ws.maxConn)
+		WebsocketError.Inc()
 		return
 	}
 
@@ -100,25 +112,29 @@ func (ws *WebSocketServer) wsHandler(w http.ResponseWriter, r *http.Request) {
 		} else {
 			slog.Error("unable to initiate ws connection", "key", ws.key, "func", "wsHandler", "err", err)
 		}
+		WebsocketError.Inc()
 		return
 	}
+
+	ActiveConnections.Inc()
+	TotalConnections.Inc()
+	ClientsPerKey.WithLabelValues(ws.key).Inc()
+	startTime := time.Now()
 
 	ws.clientsLock.Lock()
 	ws.clientConnections++
 	ws.clients = append(ws.clients, conn)
 	ws.clientsLock.Unlock()
-	defer ws.closeSocket(conn)
-
-	ws.serverStats.lock.Lock()
-	if ws.serverStats.maxNumberConn < ws.clientConnections {
-		ws.serverStats.maxNumberConn = ws.clientConnections
-	}
-	ws.serverStats.lock.Unlock()
+	defer func() {
+		ConnectionDuration.Observe(time.Since(startTime).Seconds())
+		ws.closeSocket(conn)
+	}()
 
 	ws.hardRefresh(conn)
 
 	err = ws.readLoop(conn)
 	if err != nil {
 		slog.Error("error in clients readloop", "key", ws.key, "func", "wsHandler", "err", err)
+		WebsocketError.Inc()
 	}
 }